do not expose track text to a11y api
the text on track is meant as a visual text representation of state.

However this text should *not* be exposed to the accessibility api / be included in the accessible name of the control, as it would cause redundant announcements to the change in state that is provided by the control itself (e.g. `aria-checked=“true/false”`

the addition of `aria-hidden=“true”` to the track ensures this text will not be exposed.